% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{aggregateDf}
\alias{aggregateDf}
\title{aggregateDf}
\usage{
aggregateDf(df, groupBy)
}
\arguments{
\item{df}{\code{data.frame} to aggregate}

\item{groupBy}{vector whose unique values indicates the grouping of the rows in the data frame}
}
\value{
aggregated \code{data.frame}
}
\description{
Wrapper around \code{aggregate} to merge rows of a data frame based on a specified grouping.
Merging is done by either returning the unique value for each group or if multiple different
val;ues exists converting them to character and concatenating by ";"
}
\author{
Fabian Mueller
}
