% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigeneticsData.R
\name{bed2GRanges}
\alias{bed2GRanges}
\title{bed2GRanges}
\usage{
bed2GRanges(fname, assembly = NA)
}
\arguments{
\item{assembly}{genome assembly}

\item{id}{region set id}
}
\value{
a GRanges object containing the region set
}
\description{
loads a bed file and converts it to a GRanges object. Note that bed files are 0-based, right-exclusive by definition
The output of this function will be 1-based, right inclusive as defined by GRanges.
}
\examples{
reg.fname <- system.file(file.path("extdata","deep_chip_ctrl_regions.hg19.bed"), package = "muRtools")
regs <- bed2GRanges(reg.fname)
}

