% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{bedTobigBed}
\alias{bedTobigBed}
\title{bedTobigBed}
\usage{
bedTobigBed(bedFn, chromSizes, bbFn = paste0(gsub("\\\\.bed$", "", bedFn),
  ".bb"), bedToBigBed = "bedToBigBed")
}
\arguments{
\item{bedFn}{filename of the bed file}

\item{chromSizes}{named vector of chromosome sizes}

\item{bbFn}{filename to save the bigBed file to}

\item{bedToBigBed}{executable of the 'bedToBigBed' tool}
}
\value{
nothing of particular interest
}
\description{
Convert a bed file to bigBed. requires the 'bedToBigBed' tool
}
