% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscPlots.R
\name{chordDiagramFromContingencyTable}
\alias{chordDiagramFromContingencyTable}
\title{chordDiagramFromContingencyTable}
\usage{
chordDiagramFromContingencyTable(contTab, cs_rows = colpal.mu.cat,
  cs_columns = colpal.mu.cat)
}
\arguments{
\item{contTab}{\code{matrix} or \code{table} containing the contingency matrix}

\item{cs_rows}{color scheme to use for the rows of the matrix}

\item{cs_columns}{color scheme to use for the rows of the matrix}
}
\value{
nothing of particular interest (include this function while plotting).
}
\description{
Plot a bipartite chord diagram (circular Sankey diagram with 2 categories) for a contingency matrix
the values of two matrices
}
\examples{
# contingency table of air quality quantile by month
contTab <- with(airquality, table(cut(Temp, quantile(Temp)), Month))
names(dimnames(contTab))[1] <- "quantile"
chordDiagramFromContingencyTable(contTab)
}
\author{
Fabian Mueller
}
