% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{colpal.histone}
\alias{colpal.histone}
\alias{colpal.histone,colpal.histone.ihec,colgrad.methylation.rb,colgrad.methylation.yb}
\title{Custom Color Paletes for epigenetic modifications}
\format{\preformatted{ Named chr [1:10] "#7FCDBB" "#ADDD8E" "#238B45" "#FB6A4A" ...
 - attr(*, "names")= chr [1:10] "H3K4me1" "H3K4me2" "H3K4me3" "H3K9me1" ...
}}
\usage{
colpal.histone
}
\description{
Named vectors of colors for different modifications
\describe{
  \item{\code{colpal.histone}}{
  	Histone modifications
  }
}
}
\examples{
library(gplots)
pie(rep(1,length(colpal.histone)), labels=names(colpal.histone), col=colpal.histone)
}
\keyword{datasets}

