% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{countPairwiseOverlaps}
\alias{countPairwiseOverlaps}
\title{countPairwiseOverlaps}
\usage{
countPairwiseOverlaps(grl1, grl2, ...)
}
\arguments{
\item{grl1}{list of \code{GRanges} or \code{GRangesList} object 1}

\item{grl2}{list of \code{GRanges} or \code{GRangesList} object 2}

\item{...}{arguments passed on to \code{findOverlaps}}
}
\value{
an integer matrix containing pairwise overlaps between elements in \code{grl1} and \code{grl1}
}
\description{
Fast counting of pairwise overlaps between two lists of region sets
}
