% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{diagDivCellHeatmap}
\alias{diagDivCellHeatmap}
\title{diagDivCellHeatmap}
\usage{
diagDivCellHeatmap(ml, mr, col.l = NULL, col.r = NULL,
  name.l = "Lower left", name.r = "Upper right", ...)
}
\arguments{
\item{ml}{the first value matrix (will be the left-lower diagonal in the cells of the resulting heatmap)}

\item{mr}{the second value matrix (will be the right-upper diagonal in the cells of the resulting heatmap)}

\item{col.l}{color scheme for the left-lower diagonal matrix. Should be generated by \code{circlize::colorRamp2}.
If \code{NULL} a default color scheme witll be used.}

\item{col.r}{color scheme for the upper-right diagonal matrix. Should be generated by \code{circlize::colorRamp2}.
If \code{NULL} a default color scheme witll be used.}

\item{name.l}{Name for the lower-left submatrix}

\item{name.r}{Name for the upper-right submatrix}

\item{...}{parameters passed on to \code{ComplexHeatmap::Heatmap}}
}
\value{
a \code{ComplexHeatmap::Heatmap} object containing the heatmap
}
\description{
Plot a heatmap in which each cell is subdivided into a lower-left and upper-right triangle representing
the values of two matrices
}
\examples{
m1 <- matrix(rnorm(100, mean=0), ncol=10)
m2 <- matrix(rnorm(100, mean=2), ncol=10)
rownames(m1) <- rownames(m2) <- colnames(m1) <- colnames(m2) <- paste0("Idx", 1:10)
diagDivCellHeatmap(m1, m2, cluster_rows=FALSE, cluster_columns=FALSE)
cres <- as.hclust(muRtools::getClusteringDendrogram(m1, distMethod="euclidean", linkMethod="ward.D", corMethod="pearson"))
diagDivCellHeatmap(m1, m2, cluster_rows=cres, cluster_columns=cres)
}
\author{
Fabian Mueller
}
