% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{diagDivHeatmap}
\alias{diagDivHeatmap}
\title{diagDivHeatmap}
\usage{
diagDivHeatmap(ml, mr, col.l = NULL, col.r = NULL,
  name.l = "Lower left", name.r = "Upper right", cluster = FALSE,
  cell.val.text = FALSE, cell.val.text.round = 2, ...)
}
\arguments{
\item{ml}{the first value matrix (the left-lower diagonal matrix will be in the result)}

\item{mr}{the first value matrix (the right-upper diagonal matrix will be in the result)}

\item{col.l}{color scheme for the left-lower diagonal matrix. Should be generated by \code{circlize::colorRamp2}.
   Alternatively can be a character vector specifying color levels/breaks.
If \code{NULL} a default color scheme will be used.}

\item{col.r}{color scheme for the upper-right diagonal matrix. Should be generated by \code{circlize::colorRamp2}.
Alternatively can be a character vector specifying color levels/breaks.
If \code{NULL} a default color scheme will be used.}

\item{name.l}{Name for the lower-left submatrix}

\item{name.r}{Name for the upper-right submatrix}

\item{cluster}{logical or clustering object. will be passed to the \code{cluster_rows} and \code{cluster_columns} arguments of \code{Heatmap}}

\item{cell.val.text}{logical indicating whether the cells value should be added as text}

\item{cell.val.text.round}{if the cell value is numeric, the number of digits to which the cell text is rounded}

\item{...}{parameters passed on to \code{ComplexHeatmap::Heatmap}}
}
\value{
a \code{ComplexHeatmap::Heatmap} object containing the heatmap
}
\description{
Plot a diagonally divided heatmap u
}
\examples{
m1 <- matrix(rnorm(100, mean=0), ncol=10)
m2 <- matrix(rnorm(100, mean=2), ncol=10)
rownames(m1) <- rownames(m2) <- colnames(m1) <- colnames(m2) <- paste0("Idx", 1:10)
diagDivHeatmap(m1, m2)
diagDivHeatmap(m1, m2, cell.val.text=TRUE, cell.val.text.round=3)
cres <- as.hclust(muRtools::getClusteringDendrogram(m1, distMethod="euclidean", linkMethod="ward.D", corMethod="pearson"))
diagDivHeatmap(m1, m2, cluster=cres, cell.val.text=TRUE, cell.val.text.round=2)
}
\author{
Fabian Mueller
}
