% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{dist.correlation}
\alias{dist.correlation}
\title{dist.correlation}
\usage{
dist.correlation(x, ...)
}
\arguments{
\item{x}{a matrix on which the distances should be computed}

\item{...}{parameters passed on to \code{cor()}}
}
\value{
a distance matrix
}
\description{
Compute a distance matrix based on 1-corelation
}
\author{
Fabian Mueller
}
