% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodeProjectTools.R
\name{get.encode.cell.table}
\alias{get.encode.cell.table}
\title{get.encode.cell.table}
\usage{
get.encode.cell.table(
  cvFile = "http://hgdownload.cse.ucsc.edu/goldenPath/encodeDCC/cv.ra"
)
}
\arguments{
\item{cvFile}{the file location for the ENCODE controlled vocabulary file. Defaults to the one provided by ENCODE}
}
\value{
a table containing ENCODE cell annotations
}
\description{
given the ENCODE controlled vocabulary file, retrieves a table characterizing the ENCODE cells
}
\examples{
ect <- get.encode.cell.table()
}
