% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdb.R
\name{getAnnotGrl.gencode}
\alias{getAnnotGrl.gencode}
\title{getAnnotGrl.gencode}
\usage{
getAnnotGrl.gencode(name)
}
\arguments{
\item{name}{gencode identifier. Currently supported are: "gencode.v27", "gencode.v19", "gencode.vM16", "gencode.vM1"}
}
\value{
\code{GRangesList} object with annotated elements for each element type (genes, transcripts, exons, ...)
}
\description{
Create a \code{GRangesList} with element annotation by downloading the corresponding GTF file from Gencode
}
\author{
Fabian Mueller
}
