% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{getAssocTestRes.pca}
\alias{getAssocTestRes.pca}
\title{getAssocTestRes.pca}
\usage{
getAssocTestRes.pca(X, ph, nComp = 10, nPerm = 1000)
}
\arguments{
\item{X}{A matrix on which the dimension reduction is to be performed. Alternatively, it can be a matrix of PC coordinates computed by \code{getDimRedCoords.pca}.}

\item{ph}{annotation table for the datapoints. The columns of this table will be used to test associations with the PCs.
Should be a \code{matrix} or \code{data.frame}.}

\item{nComp}{number of PCs to be considered}

\item{nPerm}{number of permutation tests to be conducted if an annotation in \code{ph} is numeric (i.e. a correlation permutation test is performed)}
}
\value{
A nested list of tested associations one element for each column in \code{ph} (1st level), each PC (2nd level).
        Each element is again a list with the name of the test being used (\code{test}), the test statistic (\code{statistic}) and p-value (\code{pvalue})
}
\description{
Test associations of annotations with principal components (PCA)
}
\author{
Fabian Mueller
}
