% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{getClusteringDendrogram}
\alias{getClusteringDendrogram}
\title{getClusteringDendrogram}
\usage{
getClusteringDendrogram(X, samplesOrdered = colnames(X), distMethod = "cor",
  linkMethod = "ward.D", corMethod = "pearson")
}
\arguments{
\item{X}{A matrix for which the sample clustering dendrogram should be computed. Samples correspond to columns and features correspond to rows.}

\item{samplesOrdered}{character vector specifying the preferred order of samples}

\item{distMethod}{distance metric to be used for clusteing. must be either "cor" or a valid distance method for \code{dist()}}

\item{linkMethod}{linkage method (see \code{hclust} for details)}

\item{corMethod}{method for computing correlation coefficients. Only relevant if \code{distMethod=="cor"}.}
}
\value{
clustering dendrogram (\code{dendrogram} object)
}
\description{
Get a clustering dendrogram using hierarchical clustering (wrapper)
}
\author{
Fabian Mueller
}
