% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisConfig.R
\name{getConfig}
\alias{getConfig}
\title{getConfig}
\usage{
getConfig(cfgFn, anaName, addDirs = TRUE)
}
\arguments{
\item{cfgFn}{JSON file name}

\item{anaName}{name of the analysis to be conducted}

\item{addDirs}{should the analysis directories be created and added to the object}
}
\value{
an S3 object containing analysis config elements
}
\description{
Get analysis configuration object from a JSON file. Automatically runs parsers
for config elements specified in the '.MU_ANA_CONFIG' section of the JSON file
}
\author{
Fabian Mueller
}
