% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{getDimRedCoords.mds}
\alias{getDimRedCoords.mds}
\title{getDimRedCoords.mds}
\usage{
getDimRedCoords.mds(X, distMethod = "euclidean")
}
\arguments{
\item{X}{A matrix on which the dimension reduction is to be performed}

\item{distMethod}{distance metric to be employed}
}
\value{
a matrix containing two columns for the reduced dimensions and the same number of
        rows as \code{X}
}
\description{
Get dimension reduction coordinates (Multidimensional Scaling)
}
\author{
Fabian Mueller
}
