% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{getDimRedCoords.pca}
\alias{getDimRedCoords.pca}
\title{getDimRedCoords.pca}
\usage{
getDimRedCoords.pca(X, components = c(1, 2), method = "prcomp", ...)
}
\arguments{
\item{X}{A matrix on which the dimension reduction is to be performed}

\item{components}{principal component to be returned}

\item{method}{Method/package to be used for computing principal components. Currently \code{prcomp} and \code{irlba} are supported.}

\item{...}{other arguments passed on to the PCA method}
}
\value{
a matrix containing two columns for the reduced dimensions and the same number of
        rows as \code{X}
}
\description{
Get dimension reduction coordinates (PCA)
}
\author{
Fabian Mueller
}
