% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{getDimRedCoords.umap}
\alias{getDimRedCoords.umap}
\title{getDimRedCoords.umap}
\usage{
getDimRedCoords.umap(X, distMethod = "euclidean", dims = c(1, 2))
}
\arguments{
\item{X}{A matrix on which the dimension reduction is to be performed}

\item{distMethod}{distance metric to be employed}

\item{dims}{dimensions to return from the reduction}
}
\value{
a matrix containing two columns for the reduced dimensions and the same number of
        rows as \code{X}
}
\description{
Get dimension reduction coordinates using the UMAP method
}
\author{
Fabian Mueller
}
