% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{getDimRedPlot}
\alias{getDimRedPlot}
\title{getDimRedPlot}
\usage{
getDimRedPlot(
  coords,
  annot = NULL,
  colorCol = NULL,
  shapeCol = NULL,
  colScheme = "[auto]",
  ptSize = 3,
  addLabels = FALSE,
  addDensity = FALSE,
  addVoronoi = FALSE,
  annot.text = NULL,
  orderCol = NULL,
  facetCols = NULL
)
}
\arguments{
\item{coords}{dimension reduction coordinates}

\item{annot}{annotation matrix with the same number of rows as \code{coord}}

\item{colorCol}{name or index in the annotation matrix (\code{annot}) that should be used for coloring the points
if \code{colorCol} not supplied but \code{annot} is supplied, it defaults to the first annotation column}

\item{shapeCol}{name or index in the annotation matrix (\code{annot}) that should be used for point shapes
if \code{shapeCol} not supplied but \code{annot} is supplied and has more than one column, 
it defaults to the second annotation column}

\item{colScheme}{color sheme to be used in coloring the points. can be a character vector with the supplied colors. Alternatively, if it is a one-element character vector \code{"[auto]"} the color scheme will be selected automatically using \code{muRtools::ggAutoColorScale}. If \code{NULL}, ggplots default color scheme will be used.}

\item{ptSize}{size of the points in the scatterplot}

\item{addLabels}{should observation labels be added to each point}

\item{addDensity}{should Gaussian Kernel density estimation be performed and the contour lines plotted for each color group}

\item{addVoronoi}{should a Voronoi tessalation grid (based on \code{colorCol}) be added to the plot}

\item{annot.text}{optional text to be added in the lower right corner of the plot}

\item{orderCol}{name or index in the annotation matrix (\code{annot}) that should be used for ordering the points. If not \code{NULL} Points will be ordered increasingly by their value, i.e. higher-valued points are plottet over lower-valued points}

\item{facetCols}{name (string) of columns to be used for faceting the resulting plot. Each facet will contain all the points not in the facet as grey points.}
}
\value{
a \code{ggplot2} object containing the dimension reduction plot
}
\description{
Generate a plot from dimension reduction coordinates
}
\examples{
\donttest{
df <- data.frame(
	x = c(rnorm(20, mean=0, sd=0.2), rnorm(10, mean=1, sd=0.4), rnorm(15, mean=1, sd=0.2)),
	y = c(rnorm(20, mean=0, sd=0.2), rnorm(10, mean=1, sd=0.4), rnorm(15, mean=0.5, sd=0.3)),
	group = rep(c("group1", "group2", "group3"), times=c(20,10,15)),
	stringsAsFactors=FALSE
)
getDimRedPlot(df[,c("x", "y")], annot=df[,c("group"), drop=FALSE], colorCol="group")
getDimRedPlot(df[,c("x", "y")], annot=df[,c("group"), drop=FALSE], colorCol="group", addDensity=TRUE)
}
}
\author{
Fabian Mueller
}
