% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdb.R
\name{getGeneAnnotMap}
\alias{getGeneAnnotMap}
\title{getGeneAnnotMap}
\usage{
getGeneAnnotMap(assembly, from = "ENSEMBL", to = "SYMBOL",
  multiMap = "paste")
}
\arguments{
\item{assembly}{character string specifying the assembly}

\item{from}{the column name that will be used as the key for the resulting map}

\item{to}{the column name that will be used as the result for the resulting map}

\item{multiMap}{character string specifying what to do if multiple mappings are found for a key
by default (\code{multiMap="paste"}) the results will be pasted into a single character string (separated by ';').
Other options include \code{'first'} for just returning the first value or 'list' for returning a list of all values}
}
\value{
a named vector (or list depending on how the \code{multiMap} argument is chosen) providing a mapping
}
\description{
Get a mapping (e.g. of identifiers) and automatically select the correct \code{AnnotationDbi} database for a given
assembly
}
\author{
Fabian Mueller
}
