% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GvizManager-class.R
\docType{methods}
\name{getGeneRegionById,GvizManager-method}
\alias{getGeneRegionById,GvizManager-method}
\alias{getGeneRegionById}
\title{getGeneRegionById-methods}
\usage{
\S4method{getGeneRegionById}{GvizManager}(object, id, offsetUp = 0.1,
  offsetDown = 0.1, type = "ensembl")
}
\arguments{
\item{object}{\code{\linkS4class{GvizManager}} object}

\item{id}{character specifying the gene id for which coordinates should be retrieved}

\item{offsetUp}{offset for retrieving a flanking region upstream of the gene. Can be an \code{integer} (e.g. \code{5L}) for the absolute number of bases
or a \code{double} that specifies the relative length of the gene}

\item{offsetDown}{offset for retrieving a flanking region downstream of the gene. Can be an \code{integer} (e.g. \code{5L}) for the absolute number of bases
or a \code{double} that specifies the relative length of the gene}

\item{type}{type of identifier. currently only \code{'ensembl'} and \code{'ensembl_long'} are supported}
}
\value{
a \code{GRanges} object containing the gene coordinates
}
\description{
Retrieve the coordinates of a gene given the gene identifier
}
\author{
Fabian Mueller
}
