% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GvizManager-class.R
\docType{methods}
\name{getGeneRegionBySymbol,GvizManager-method}
\alias{getGeneRegionBySymbol,GvizManager-method}
\alias{getGeneRegionBySymbol}
\title{getGeneRegionBySymbol-methods}
\usage{
\S4method{getGeneRegionBySymbol}{GvizManager}(object, symbol,
  offsetUp = 0.1, offsetDown = 0.1, exactMatch = TRUE,
  ignore.case = FALSE)
}
\arguments{
\item{object}{\code{\linkS4class{GvizManager}} object}

\item{symbol}{character specifying the gene symbol for which coordinates should be retrieved}

\item{offsetUp}{offset for retrieving a flanking region upstream of the gene. Can be an \code{integer} (e.g. \code{5L}) for the absolute number of bases
or a \code{double} that specifies the relative length of the gene}

\item{offsetDown}{offset for retrieving a flanking region downstream of the gene. Can be an \code{integer} (e.g. \code{5L}) for the absolute number of bases
or a \code{double} that specifies the relative length of the gene}

\item{exactMatch}{should the symbol match exactly or should it just be contained in the symbol annotation}

\item{ignore.case}{ignore the case when matching the symbol}
}
\value{
a \code{GRanges} object containing the gene coordinates
}
\description{
Retrieve the coordinates of a gene given the gene symbol
}
\author{
Fabian Mueller
}
