% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GvizManager-class.R
\docType{methods}
\name{getGeneTrack,GvizManager-method}
\alias{getGeneTrack,GvizManager-method}
\alias{getGeneTrack}
\title{getGeneTrack-methods}
\usage{
\S4method{getGeneTrack}{GvizManager}(object, chrom)
}
\arguments{
\item{object}{\code{\linkS4class{GvizManager}} object}

\item{chrom}{name of the chromosome for which the track should be returned}
}
\value{
\code{Gviz::GeneRegionTrack} for the specified chromosome
}
\description{
Get a \code{Gviz} gene annotation track for a specified chromosome
}
\author{
Fabian Mueller
}
