% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{getGenomeGr}
\alias{getGenomeGr}
\title{getGenomeGr}
\usage{
getGenomeGr(assembly, ...)
}
\arguments{
\item{assembly}{assembly}

\item{...}{other arguments passed on to \code{setGenomeProps}}
}
\value{
\code{GRanges} object
}
\description{
retrieve the full genome as GRanges object
}
