% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{getGenomeObject}
\alias{getGenomeObject}
\title{getGenomeObject}
\usage{
getGenomeObject(assembly)
}
\arguments{
\item{assembly}{string specifying the assembly}
}
\value{
\code{BSgenome} object
}
\description{
retrieve the appropriate \code{BSgenome} for an assembly string
}
