% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epigeneticsData.R
\name{getRegionSet}
\alias{getRegionSet}
\title{getRegionSet}
\usage{
getRegionSet(id, assembly)
}
\arguments{
\item{id}{region set id}

\item{assembly}{genome assembly}
}
\value{
a GRanges object containing the region set
}
\description{
retrieves custom region sets by id
}
\examples{
regs <- getRegionSet("deep_chip_ctrl_regions","hg19")
}
