% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{getSeqlengths4assembly}
\alias{getSeqlengths4assembly}
\title{getSeqlengths4assembly}
\usage{
getSeqlengths4assembly(assembly)
}
\arguments{
\item{assembly}{assembly}
}
\value{
named vector of chromosomes/contigs and sequence lengths
}
\description{
retrieve chromosomes/contigs and their sequence lengths for known assemblies
}
