% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txdb.R
\name{getTxDb.gencode}
\alias{getTxDb.gencode}
\title{getTxDb.gencode}
\usage{
getTxDb.gencode(name)
}
\arguments{
\item{name}{gencode identifier. Currently supported are: "gencode_v27", "gencode_v19", "gencode_vM16", "gencode_vM1"}
}
\value{
\code{TxDb} object
}
\description{
Create a \code{TxDb} object by downloading the corresponding GTF file from Gencode
}
\author{
Fabian Mueller
}
