% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{ggAutoColorScale}
\alias{ggAutoColorScale}
\title{ggAutoColorScale}
\usage{
ggAutoColorScale(x, method = "color", symmetric = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{method}{method for scaling: \code{"color"} or \code{"fill"}}

\item{symmetric}{treat numeric values as symmetric. If there are values smaller and larger than 0, a diverging color scheme will be applied}
}
\value{
the theme structure
}
\description{
Automatical color scales for values for ggplots
}
\examples{
dframe.num.pos <- data.frame(x=runif(100),y=runif(100))
ggplot(dframe.num.pos, aes(x=x,y=y, color=x)) + geom_point() + ggAutoColorScale(dframe.num.pos[,"x"])
dframe.num.sym <- data.frame(x=rnorm(100),y=rnorm(100))
ggplot(dframe.num.sym, aes(x=x,y=y, color=x)) + geom_point() + ggAutoColorScale(dframe.num.sym[,"x"])
dframe.num.sym.lab <- data.frame(x=rnorm(100),y=rnorm(100), lab=sample(c("A", "B", "C", "D"), 100, replace=TRUE))
ggplot(dframe.num.sym.lab, aes(x=x,y=y, color=lab)) + geom_point() + ggAutoColorScale(dframe.num.sym.lab[,"lab"])
}
