% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{ggMessagePlot}
\alias{ggMessagePlot}
\title{ggMessagePlot}
\usage{
ggMessagePlot(txt)
}
\arguments{
\item{txt}{Text to be plotted.}
}
\value{
The newly initialized \code{ggplot} instance.
}
\description{
Creates a plot, using \pkg{ggplot2}, with a single text message.
}
\examples{
\donttest{
ggMessagePlot("Missing data")
}
}
