% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{ggplot2.distr}
\alias{ggplot2.distr}
\title{ggplot2.distr}
\usage{
ggplot2.distr(x, fillColor = "#676D8D")
}
\arguments{
\item{x}{vector of values whose distribution is to be plottet}

\item{fillColor}{color to be used to fill the violin}
}
\value{
the ggplot2 object (can be extended for plotting)
}
\description{
Distribution plot combining violin and boxplot using ggplot2
}
\examples{
x <- rnorm(1000)
ggplot2.distr(x)
}
