% Generated by roxygen2 (4.0.1): do not edit by hand
\name{ggplot2.heatmap}
\alias{ggplot2.heatmap}
\title{ggplot2.heatmap}
\usage{
ggplot2.heatmap(mm, add.text = FALSE)
}
\arguments{
\item{mm}{matrix or dataframe to be plottet as heatmap}

\item{add.text}{logical. should the cells be labelled with the values}
}
\value{
the ggplot2 object (can be extended for plotting)
}
\description{
converts a matrix or dataframe into a ggplot2 object for subsequent plotting.
}
\examples{
ggplot2.heatmap(airquality[1:15,],add.text=TRUE) + scale_fill_gradient(low = "red",high = "steelblue"))
}

