% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{ggsave4doc}
\alias{ggsave4doc}
\title{ggsave4doc}
\usage{
ggsave4doc(
  fn,
  plot = last_plot(),
  width = 192,
  height = 192,
  units = "mm",
  family = "Helvetica",
  dimPreset = NULL,
  useDingbats = FALSE,
  ...
)
}
\arguments{
\item{fn}{file name}

\item{plot}{see \code{?ggsave}}

\item{width}{see \code{?ggsave}}

\item{height}{see \code{?ggsave}}

\item{units}{see \code{?ggsave}}

\item{family}{see \code{?ggsave}}

\item{dimPreset}{presets for figure dimensions. Possible values are \code{NULL} (don't use a preset; default),
\code{"slide_nuc_wide_full"} (full slide using Fabian's wide nucleosome template),
\code{"slide_nuc_wide_half"} (half a slide using Fabian's wide nucleosome template)
Overwrites \code{width}, \code{height} and \code{units}.}

\item{useDingbats}{see \code{?ggsave}}

\item{...}{see \code{?ggsave}}
}
\value{
result of \code{ggsave} command
}
\description{
Wrapper around ggsave that has default values for parameters fitting for embedding
plots into my documents
}
