% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{ggtemp}
\alias{ggtemp}
\title{ggtemp}
\usage{
ggtemp(
  plot = last_plot(),
  fn = paste0("~/tmp_work/", getHashString("ggplot"), ".pdf"),
  ...
)
}
\arguments{
\item{plot}{see \code{?ggsave}}

\item{fn}{file name}

\item{...}{see \code{?ggsave}}
}
\value{
result of \code{ggsave} command
}
\description{
Wrapper for quickly saving plot to temporary file
}
