% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichmentTools.R
\name{goEnrichment}
\alias{goEnrichment}
\title{goEnrichment}
\usage{
goEnrichment(
  qids,
  uids,
  ontology = "BP",
  idType = "ensembl",
  assembly = "org.Hs.eg.db",
  algorithm = "weight01"
)
}
\arguments{
\item{qids}{character vector of query gene IDs}

\item{uids}{character vector of universe gene IDs}

\item{ontology}{character specifying the ontology to use (default: \code{"BP"})}

\item{idType}{character specifying which universe the gene IDs come from (default: \code{"ensembl"}). Possible values are: \code{"entrez", "genbank", "alias", "ensembl", "symbol", "genename", "unigene"}}

\item{assembly}{character specifying the genome to use. Can either be the name of the package to be used for mapping the identifiers (e.g. \code{"org.Hs.eg.db"}; default)
or an identifier for a genome assembly ((e.g. \code{"hg38"})}

\item{algorithm}{algorithm employed by \code{topGO}. See \code{topgGO::runTest} for details.}
}
\value{
a list (S3 class) object containing:
        - \code{$tgData}:    the used \code{topGOdata} object
        - \code{$resultObj}: the resulting \code{topGOresult} object
        - \code{$table}:     a summary table of statistics for each GO term
}
\description{
Perform Gene Ontology (GO) enrichment using the \code{topGO} package
}
\author{
Fabian Mueller
}
