% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{granges2bed.igv}
\alias{granges2bed.igv}
\title{granges2bed.igv}
\usage{
granges2bed.igv(gr, fn, trackName = NULL, scoreCol = NULL,
  na.rm = FALSE, nameCol = NULL, col.cat = colpal.bde,
  col.cont = c("#EDF8B1", "#41B6C4", "#081D58"), col.na = "#bdbdbd",
  col.range = NULL, doSort = TRUE)
}
\arguments{
\item{gr}{GRanges object}

\item{fn}{filename to save bed file to}

\item{trackName}{track name to be displayed}

\item{scoreCol}{the score column (in the GRanges elementMetadata) that is optionally used for coloring}

\item{na.rm}{flag indicating whether items with NA score should be removed}

\item{nameCol}{the name column (in the GRanges elementMetadata) that is used for labelling the items}

\item{col.cat}{color panel for coloring categorical scores}

\item{col.cont}{color panel for coloring numerical scores}

\item{col.na}{color used for NA scores}

\item{col.range}{vector of length 2 indicating the range of scores for the color scales to be applied (continuous scores only)}

\item{doSort}{sort the regions before writing the output}

\item{na.rm}{flag indicating whether items with NA score should be removed}
}
\value{
invisibly, the resulting data frame containing the bed file columns
}
\description{
Save a GRanges object to a bed file which can be displayed by IGV
}
