% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{granges2bed}
\alias{granges2bed}
\title{granges2bed}
\usage{
granges2bed(gr, fn, score = NULL, addAnnotCols = FALSE, colNames = FALSE,
  doSort = TRUE)
}
\arguments{
\item{gr}{GRanges object}

\item{fn}{filename to save bed file to}

\item{addAnnotCols}{add the columns stored in elementMetadata of GRanges}

\item{colNames}{add column names}

\item{doSort}{sort the regions before writing the output}

\item{sc}{score vector or column in elementMetadata of GRanges}
}
\value{
result of writing the table (see \code{write.table})
}
\description{
Save a GRanges object to a bed file
}

