% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{granges2igv}
\alias{granges2igv}
\title{granges2igv}
\usage{
granges2igv(gr, fn, addStrand = FALSE, addAnnotCols = TRUE, doSort = TRUE,
  toTDF = FALSE)
}
\arguments{
\item{gr}{GRanges object}

\item{fn}{filename to save IGV file to}

\item{addAnnotCols}{add the columns stored in elementMetadata of GRanges}

\item{doSort}{sort the regions before writing the output}

\item{toTDF}{convert to TDF file. Requires that "igvtools" is executable from the current path}

\item{sc}{score vector or column in elementMetadata of GRanges}
}
\value{
result of writing the table (see \code{write.table})
}
\description{
Save a GRanges object to a IGV file
}

