% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{grLiftOver}
\alias{grLiftOver}
\title{grLiftOver}
\usage{
grLiftOver(gr, targetAssembly, onlyUnique = TRUE)
}
\arguments{
\item{gr}{\code{GRanges} object to liftOver}

\item{targetAssembly}{character string specifying the target assembly}
}
\value{
\code{GRanges} object with coordinates that could uniquely be
}
\description{
Converts coordinates of a GRanges object to target genome assembly. Wraps around rtracklayer::liftOver
and automatically downloads and selects the correct chain file
}
