% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{grTile}
\alias{grTile}
\title{grTile}
\usage{
grTile(gr, tile.width = 200, keepMetadata = TRUE)
}
\arguments{
\item{gr}{\code{GRanges} object to liftOver}

\item{tile.width}{length of the tiling window}

\item{keepMetadata}{Should the metadata columns for each element be preserved in the resulting object}
}
\value{
\code{GRanges} containing the tiling regions. Additional metadata columns named \code{.orgIdx}, \code{.winIdx} denote the indices
        of the original element and the window respectively
}
\description{
Tile each element in a \code{GRanges} object into equally-sized windows.
If the length of an element is not divisible by the window-size, each element will be adjusted to match a multiple of the desired window-size
}
