% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_gViz_ext.R
\name{gviz_bwScheme}
\alias{gviz_bwScheme}
\title{gviz_bwScheme}
\usage{
gviz_bwScheme()
}
\value{
(invisibly) the scheme structure
}
\description{
A Gviz scheme with black labels and transparent boxes for labels
}
