% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GvizManager-class.R
\docType{class}
\name{GvizManager-class}
\alias{GvizManager-class}
\title{GvizManager}
\description{
A class for storing information on pipeline jobs.
}
\section{Slots}{

\describe{
  \item{\code{assembly}}{
	Genome assembly as string
  }
  \item{\code{txdb}}{
	\code{TxDb} object for the gene model
  }
  \item{\code{geneAnnot}}{
	\code{GRanges} object with gene annotation (such as gene identifiers and symbols)
  }
}
}

\section{Methods}{

\describe{
   \item{\code{\link{firstMethod,GvizManager-method}}}{
     Description of the first method
   }
}
}

\author{
Fabian Mueller
}
