% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listOps.R
\name{indicesInList}
\alias{indicesInList}
\title{indicesInList}
\usage{
indicesInList(x, l)
}
\arguments{
\item{x}{vector of items to be found in the list}

\item{l}{list of vectors in which \code{x} should be found}
}
\value{
a list containing an element for each item in \code{x} that contains the indices of its occurrence in \code{l}
}
\description{
Find the occurrences of items in a list of vectors
}
\examples{
l <- list(1:3, 4:5, 5:9)
x <- c(2,3,5,666,8,5)
indicesInList(x, l)
}

