% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muKDEplots.R
\name{kde.plot.simple}
\alias{kde.plot.simple}
\alias{kde.plot}
\alias{kde.plot.matrix}
\title{kde.plot}
\usage{
kde.plot.simple(x, initial = FALSE, col = c("#00640044"), ...)

kde.plot.matrix(X, col = makeTrans(rainbow(ncol(X))), legend = TRUE,
  ...)
}
\arguments{
\item{x}{values}

\item{initial}{is the plot initial, i.e. should the plot() function be called}

\item{col}{color(s)}

\item{...}{more graphical parameters}

\item{X}{matrix}

\item{legend}{add a color legend with colnames of the matrix}
}
\value{
Nothing particularly interesting
}
\description{
plot a single vector as kernel density estimation (kde.plot.simple), plot columns of a matrix as kernel density estimation (kde.plot.matrix)
}
\examples{
kde.plot.simple(c(1:10,rep(5,3)),initial=TRUE,col=c("#00640044"),main="Some plot")
kde.plot.simple(rep(8,2),initial=FALSE,col=c("#64640044"))

dd <- USJudgeRatings
dd[,5] <- 1/dd[,5]
tt <- dd[,1:5]
kde.plot.matrix(tt)
}
