% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnbeads.R
\name{loadRnBeadsAnalysis}
\alias{loadRnBeadsAnalysis}
\title{loadRnBeadsAnalysis}
\usage{
loadRnBeadsAnalysis(
  input,
  type = "cluster_run",
  preprocessed = TRUE,
  setOptions = TRUE,
  preAnalysis = TRUE,
  diffMeth = TRUE
)
}
\arguments{
\item{input}{input directory}

\item{type}{analysis type. Currently only "cluster_run" is supported}

\item{preprocessed}{determines the type of RnBSet object to be loaded. If \code{TRUE} (default)
the preprossed/filtered object will be loaded. Otherwise the raw, imported object}

\item{setOptions}{should the analysis options be set from the options settings of the RnBeads run?}

\item{preAnalysis}{should the corresponding preanalysis script be called}

\item{diffMeth}{should the differential analysis result be loaded in addition?}
}
\value{
A \code{list} containing the \code{RnBSet} object (list item \code{rnbs}) and optionally
        the differential methylaiton result (list item \code{diffmeth})
}
\description{
Loads RnBeads analysis results (RnBSet object, differential methylation) and optionally
runs the corresponding preanalysis script and sets options
}
\author{
Fabian Mueller
}
