% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolaUtils.R
\name{lolaRegionSetHeatmap}
\alias{lolaRegionSetHeatmap}
\title{lolaRegionSetHeatmap
Plot a heatmap in which the rows are different user sets and the color corresponds to an enrichment score}
\usage{
lolaRegionSetHeatmap(lolaDb, lolaRes, scoreCol = "pValueLog",
  orderCol = scoreCol, signifCol = "qValue", markSignif = FALSE,
  includedCollections = c(), recalc = TRUE, pvalCut = 0.01,
  maxTerms = 50, colorpanel = colpal.cont(9, "cb.OrRd"),
  groupByCollection = TRUE, orderDecreasing = NULL)
}
\arguments{
\item{lolaDb}{LOLA DB object as returned by \code{LOLA::loadRegionDB} or \link{\code{loadLolaDbs}}}

\item{lolaRes}{LOLA enrichment result as returned by the \code{runLOLA} function from the \code{LOLA} package}

\item{scoreCol}{column name in \code{lolaRes} to be plotted}

\item{orderCol}{column name in \code{lolaRes} which is used for sorting the results}

\item{signifCol}{column name of the significance score in \code{lolaRes}. Should be one of \code{c("pValueLog", "qValue")}}

\item{markSignif}{mark significant enrichments in the heatmap}

\item{includedCollections}{vector of collection names to be included in the plot. If empty (default), all collections are used}

\item{recalc}{recalculate adjusted p-value/q-value and ranks after the specified subsetting (by \code{includedCollections})}

\item{pvalCut}{p-value cutoff (negative log10) to be employed for filtering the results}

\item{maxTerms}{maximum number of items to be included in the plot}

\item{colorpanel}{colorpanel for the heatmap gradient}

\item{groupByCollection}{facet the plot by collection}

\item{orderDecreasing}{flag indicating whether the value in \code{orderCol} should be considered as decreasing (as opposed
to increasing). \code{NULL} (default) for automatic determination.}
}
\value{
ggplot object containing the plot
}
\description{
lolaRegionSetHeatmap
Plot a heatmap in which the rows are different user sets and the color corresponds to an enrichment score
}
\author{
Fabian Mueller
}
