% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{matchStrand}
\alias{matchStrand}
\title{matchStrand}
\usage{
matchStrand(values)
}
\arguments{
\item{values}{character vector or factor of strand names}
}
\value{
Factor of genomic strand (with levels \code{"+", "-", "*"})
}
\description{
match commonly used strand names to \code{"+", "-", "*"}
}
