% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizePercentile}
\alias{normalizePercentile}
\title{normalizePercentile}
\usage{
normalizePercentile(X)
}
\value{
a matrix containing the normalized values
}
\description{
Performs percentile normalization on the columns of a matrix,
i.e. each element in a column will be the percentile it lies in in its column
}
\author{
Fabian Mueller
}
