% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalizeRank}
\alias{normalizeRank}
\title{normalizeRank}
\usage{
normalizeRank(X, out = "percentile", ties.method = "average")
}
\arguments{
\item{X}{A matrix which should be normalized}

\item{out}{output type. Either \code{"rank"} or \code{"percentile"}}

\item{ties.method}{method for breaking ties (see \code{?colRanks} for details)}
}
\value{
a matrix containing the normalized avlues
}
\description{
Performs rank normalization on the columns of a matrix
}
\author{
Fabian Mueller
}
