% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muPairs.R
\name{panel.density}
\alias{panel.density}
\title{panel.density}
\usage{
panel.density(x, y, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{...}{more plotting parameters}
}
\value{
Nothing particularly interesting
}
\description{
creates a scatterplot density panel.
}

