% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commandLine.R
\name{parse.cl.args}
\alias{parse.cl.args}
\title{parse.cl.args}
\usage{
parse.cl.args()
}
\value{
a named list with command line arguments
}
\description{
parser for command line arguments
}
\examples{
cmd.args <- parse.cl.args()
}

