% Generated by roxygen2 (4.0.1): do not edit by hand
\name{parse.encode.cv.file}
\alias{parse.encode.cv.file}
\title{parse.encode.cv.file}
\usage{

  parse.encode.cv.file(cvFile = "http://hgdownload.cse.ucsc.edu/goldenPath/encodeDCC/cv.ra")
}
\arguments{
\item{cvFile}{the file location for the ENCODE controlled vocabulary file. Defaults to the one provided by ENCODE}
}
\value{
a list containing blank line seperated blocks in each element. Each element
        is a named list containing the content to each keyword. a keyword is the first word in a line.
}
\description{
parser for the ENCODE projects controlled vocabulary file
}
\examples{
cv.blocks <- parse.cv.file()
}

