% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{plot.colpal}
\alias{plot.colpal}
\title{plot.colpal}
\usage{
\method{plot}{colpal}(cp, type = "pie")
}
\arguments{
\item{cp}{color palette, i.e. vector of colors}

\item{type}{pie chart or stripes}
}
\value{
nothing of particular interest
}
\description{
Get a continuous color palette
}
\author{
Fabian Mueller
}
