% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{plotAllDimRed}
\alias{plotAllDimRed}
\title{plotAllDimRed}
\usage{
plotAllDimRed(
  X,
  fn.prefix = NULL,
  fn.suffix = "",
  annot = NULL,
  distMethods = c(euc = "euclidean", man = "manhattan"),
  width = 10,
  height = 10,
  ...
)
}
\arguments{
\item{X}{feature matrix containing one row for each observation and one column for each feature}

\item{fn.prefix}{file prefix to be used for the resulting plots. 
If \code{NULL}, no plot is actually created, but a list of resulting plot objects is returned}

\item{fn.suffix}{file suffix to be used for the resulting plots}

\item{annot}{annotation matrix with the same number of rows as \code{X}}

\item{distMethods}{distance methods for MDS and t-SNE}

\item{width}{width of the resulting plot}

\item{height}{height of the resulting plot}

\item{...}{arguments to be passed on to \code{getDimRedPlot}}
}
\value{
(invisibly) a list of lists containing the created plots as ggplot objects and additional info for each plot
}
\description{
Generate a plots with multiple methods and parameter settings from a feature matrix
}
\details{
Currently, PCA, MDS and t-SNE are employed by default with euclidean and manhattan distance metrics where applicable
}
\author{
Fabian Mueller
}
