% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{plotCorPhm}
\alias{plotCorPhm}
\title{plotCorPhm}
\usage{
plotCorPhm(cc, clustDend = NULL, sampleAnnot = NA,
  color = colorRampPalette(rev(brewer.pal(n = 11, name = "RdBu")))(100),
  breaks = seq(-1 - 1e-06, 1 + 1e-06, length.out = length(color)),
  border_color = NA, ...)
}
\arguments{
\item{cc}{a correlation matrix (as returned by \code{cor()})}

\item{clustDend}{a clustering dendrogram to be used. Set to \code{NULL} to disable clustering dendrogram}

\item{sampleAnnot}{a data.frame containing sample information to color by (corrsponds to \code{annotation_row} and \code{annotation_col} parameters of \code{pheatmap()})}

\item{color}{see \code{?pheatmap} for details}

\item{breaks}{see \code{?pheatmap} for details. In this wrapper, the default value corresponds to splitting color across the full range of correlation coefficients [-1,1]}

\item{border_color}{see \code{?pheatmap} for details}

\item{...}{parameters passed on to \code{pheatmap()}}
}
\value{
invisibly the result of a call to \code{pheatmap()}
}
\description{
Plot a correlation matrix as heatmap. Wraps around \code{pheatmap}. Note that no clustering will be performed if not supplied with an appropriate clustering dendrogram
}
\author{
Fabian Mueller
}
