% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{plotDimRed}
\alias{plotDimRed}
\title{getDimRedPlot}
\usage{
plotDimRed(X, dimRedFun = getDimRedCoords.pca, ...)
}
\arguments{
\item{X}{feature matrix containing one row for each observation and one column for each feature}

\item{dimRedFun}{function to do dimension reduction. E.g. \code{getDimRedCoords.pca}, \code{getDimRedCoords.mds}, \code{getDimRedCoords.tsne},}

\item{...}{arguments to be passed on to \code{getDimRedPlot}}
}
\value{
a \code{ggplot2} object containing the dimension reduction plot
}
\description{
Generate a plot from a feature matrix
}
\author{
Fabian Mueller
}

