% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{plotFisherTest}
\alias{plotFisherTest}
\title{plotFisherTest}
\usage{
plotFisherTest(x, y, name.x = NULL, name.y = NULL, ...)
}
\arguments{
\item{x}{factor object or one that can be coerced to one}

\item{y}{factor object or one that can be coerced to one}

\item{name.x}{optional character string specifying the name for the first grouping}

\item{name.y}{optional character string specifying the name for the second grouping}

\item{...}{arguments passed on to \code{fisher.test}}
}
\value{
an \code{S3} object containing the test result object as returned by \code{fisher.test} and a \code{ggplot} object
}
\description{
Conduct a Fisher's exact test and plot the results as a heatmap
}
\author{
Fabian Mueller
}
