% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{plottemp}
\alias{plottemp}
\alias{ggtemp}
\title{plottemp}
\usage{
plottemp(fn = paste0("~/tmp_work/", getHashString("rplot"), ".pdf"), ...)
}
\arguments{
\item{fn}{file name}

\item{...}{see \code{?pdf}}
}
\value{
nothing of particular interest
}
\description{
Wrapper for quickly saving plot to temporary pdf file. terminate using \code{off()}
}
