% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{pngtemp}
\alias{pngtemp}
\title{pngtemp}
\usage{
pngtemp(
  fn = paste0("~/tmp_work/", getHashString("rplot"), ".png"),
  width = 1024,
  height = 1024,
  ...
)
}
\arguments{
\item{fn}{file name}

\item{...}{see \code{?png}}
}
\value{
nothing of particular interest
}
\description{
Wrapper for quickly saving plot to temporary png file. terminate using \code{dev.off()}
}
