% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{randomGroupedHeatmap}
\alias{randomGroupedHeatmap}
\title{randomGroupedHeatmap}
\usage{
randomGroupedHeatmap(
  n.row = 20,
  n.col = 6,
  ngrps.row = 2,
  ngrps.col = 3,
  cols = NULL,
  ...
)
}
\arguments{
\item{n.row}{number of rows}

\item{n.col}{number of columns}

\item{ngrps.row}{number of groups to group rows into}

\item{ngrps.col}{number of groups to group columns into}

\item{cols}{color scheme. Should be a color character vector.
If \code{NULL} a default color scheme will be used.}

\item{...}{parameters passed on to \code{ComplexHeatmap::Heatmap}}
}
\value{
a \code{ComplexHeatmap::Heatmap} object containing the heatmap
}
\description{
generate a random grouped heatmap using \code{ComplexHeatmap}.
}
\examples{
randomGroupedHeatmap(n.row=10, n.col=3, ngrps.row=2, ngrps.col=3, cols=colpal.cont(n=9, name="viridis"))
randomGroupedHeatmap(n.row=100, n.col=18, ngrps.row=3, ngrps.col=3, cols=colpal.cont(n=9, name="cb.BrBG"))
pdftemp()
draw(randomGroupedHeatmap(n.row=100, n.col=6, ngrps.row=3, ngrps.col=3, cols=colpal.PhFr.a))
dev.off()
}
\author{
Fabian Mueller
}
