% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{readTab}
\alias{readTab}
\title{readTab}
\usage{
readTab(
  fn,
  sep = "\\t",
  header = TRUE,
  stringsAsFactors = FALSE,
  quote = "",
  comment.char = "",
  na.strings = "",
  ...
)
}
\arguments{
\item{fn}{filename to read}

\item{sep}{see \code{?read.table}}

\item{header}{see \code{?read.table}}

\item{stringsAsFactors}{see \code{?read.table}}

\item{quote}{see \code{?read.table}}

\item{comment.char}{see \code{?read.table}}

\item{...}{passed to \code{read.table}}
}
\value{
the result of \code{read.table}
}
\description{
Wrapper around \code{read.table} to read tab-separated tables by default
}
\author{
Fabian Mueller
}
