% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reloadPackage}
\alias{reloadPackage}
\title{reloadPackage}
\usage{
reloadPackage(package.name)
}
\arguments{
\item{package.name}{name (string) of the package to be reloaded}
}
\value{
result of \code{library(package.name)}
}
\description{
reloads a package without quitting the R session. Useful for developing packages interactively
}
\examples{
library(GenomicRanges)
reloadPackage("GenomicRanges")
}

