% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{rowTtest}
\alias{rowTtest}
\title{rowTtest}
\usage{
rowTtest(X, idx1, idx2 = -idx1, na.rm = FALSE, alternative = "two.sided")
}
\arguments{
\item{X}{Matrix on which the test is performed for every row}

\item{idx1}{column indices of group 1 members}

\item{idx2}{column indices of group 2 members}

\item{na.rm}{Should NAs be removed (logical)}

\item{alternative}{Testing alternative. Must be one of "two.sided" (default),"less","greater" or "all".
in case of "all" a data frome with corresping alternative variables is returned. 
Otherwise the result is a vector.}
}
\value{
vector (or data.frame if alternative=="all") of p-values resulting from the Welch's t-test
}
\description{
performs a two-sided Welch's t-test (unequal variances, equal or unequal sample sizes) on each row of a matrix X with the indices inds.1 vs indices idx2 as group assignments.
}
\note{
Requires \code{matrixStats} package
}
\author{
Fabian Mueller
}
