% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lolaUtils.R
\name{runLOLA_list}
\alias{runLOLA_list}
\title{runLOLA_list
More robust version of lola for lists of user sets
(to avoid "Negative c entry in table" errors)}
\usage{
runLOLA_list(userSets, userUniverse, lolaDb, ...)
}
\arguments{
\item{userSets}{NAMED list or GRangesList of user sets}

\item{userUniverse}{GRanges object to be used as universe (as required by \code{LOLA::runLOLA})}

\item{lolaDb}{LOLA DB object as returned by \code{LOLA::loadRegionDB} or \code{loadLolaDbs}}

\item{...}{other arguments to iterate over}
}
\value{
LOLA result as returned by \code{LOLA::runLOLA}
}
\description{
runLOLA_list
More robust version of lola for lists of user sets
(to avoid "Negative c entry in table" errors)
}
\author{
Fabian Mueller
}
