% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scatter.twogroups}
\alias{scatter.twogroups}
\title{scatter.twogroups}
\usage{
scatter.twogroups(x, y, is.g, cols.all = blues9[-(1:3)], cols.g = c("coral",
  "darkred"), ...)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}

\item{is.g}{logical vector indicating whether a point is in the highlighted group}

\item{cols.all}{colorscheme for all points}

\item{cols.g}{colorscheme for the points in the highlighted group}

\item{...}{more plotting parameters}

\item{x}{x coordinates}
}
\value{
Nothing particularly interesting
}
\description{
density scatterplot with highlighting the points of a certain group in a different color
}

