% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{setGenomeProps}
\alias{setGenomeProps}
\title{setGenomeProps}
\usage{
setGenomeProps(gr, assembly, stripChrFromGenome = TRUE, ...)
}
\arguments{
\item{gr}{GRanges object to modify}

\item{assembly}{assembly}

\item{stripChrFromGenome}{Tflag indicating whether to strip away the "_chr" suffix from the genome name in the result}

\item{...}{arguments passed on to \code{getSeqlengths4assembly}}
}
\value{
GRanges object with genome properties set
}
\description{
Set the genome properties for a GRanges object given the name of a genome assembly
}
