% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{sortGr}
\alias{sortGr}
\title{sortGr}
\usage{
sortGr(gr, alnum = FALSE)
}
\arguments{
\item{gr}{\code{GRanges} object to sort}

\item{alnum}{sort chromosomes alphanumerically instead of by number}
}
\value{
sorted \code{GRanges} object
}
\description{
sort a \code{GRanges} object
}
