% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOps.R
\name{strTrim}
\alias{strTrim}
\title{strTrim}
\usage{
strTrim(ss, len.out = 50, trim.str = "...",
  len.pref = ceiling((len.out - nchar(trim.str))/2), len.suf = len.out
  - len.pref - nchar(trim.str))
}
\arguments{
\item{ss}{character vector}

\item{len.out}{target output length to trim to}

\item{trim.str}{string to place in between prefix and suffix}

\item{len.pref}{length of the prefix to be used from the original string}

\item{len.suf}{length of the suffix to be used from the original string}
}
\value{
character vector in which each element has length\code{<=len.out}
}
\description{
trim a character vector to have a desired length by taking the beginning of the string
and the end of the string
}
\author{
Fabian Mueller
}
