% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{summarizeSetOverlap}
\alias{summarizeSetOverlap}
\title{summarizeSetOverlap}
\usage{
summarizeSetOverlap(
  set1,
  set2,
  set1name = "set1",
  set2name = "set2",
  doVenn = TRUE
)
}
\arguments{
\item{set1}{vector containing elements in set 1}

\item{set2}{vector containing elements in set 2}

\item{set1name}{name for set 1}

\item{set2name}{name for set 2}

\item{doVenn}{plot a Venn diagram}
}
\value{
nothing of interest. If \code{doVenn}, a venn diagram will be plotted to the current plotting device
}
\description{
prints overlap statistics for two sets
}
\examples{
summarizeSetOverlap(1:50, 23:100, "1:50", "23:100")
}
\author{
Fabian Mueller
}
