% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTest.R
\name{testAssoc}
\alias{testAssoc}
\title{Tests for association between two vectors. Based on \code{RnBeads:::test.traits}}
\usage{
testAssoc(x, y, permMat = NULL)
}
\arguments{
\item{x}{Sample values for the first trait. This must be a vector of type \code{factor}, \code{integer} or
\code{numeric}.}

\item{y}{Sample values for the second trait. This must be a vector of type \code{factor}, \code{integer} or
\code{numeric}.}

\item{permMat}{Matrix of sample permutations (indices of x that will be used in permutation tests) in case none of the traits is a \code{factor}, and thus
permutation-based p-value from correlations is computed. If this parameter is \code{NULL} and
both \code{x} and \code{y} are sequences of numbers, no p-value is calculated.}
}
\value{
List of four elements:
                   \describe{
                     \item{error}{Error, if any, that prevented this function from computing a p-value for trait
                          association.}
                     \item{test}{Type of test performed. This is one of \code{"Fisher"}, \code{"Wilcoxon"},
                          \code{"Kruskal-Wallis"}, \code{"Correlation"} or \code{NA}. The last value indicates that
                          the traits cannot be tested for association.}
                     \item{correlation}{Value of the pearson correlation coefficient between \code{x} and \code{y},
                          or \code{NA} if any of them is \code{factor}.}
                     \item{pvalue}{Calculated p-value, or \code{NA} if the traits cannot be tested for association.}
                   }
}
\description{
Tests for association between two vectors. Based on \code{RnBeads:::test.traits}
}
