% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{textSearch}
\alias{textSearch}
\title{textSearch}
\usage{
textSearch(s, x, ...)
}
\arguments{
\item{s}{string or expression}

\item{x}{string vector to search in}
}
\value{
string vector of matches
}
\description{
Shortcut wrapper around \code{aggregate} to search case insensive in a vector of strings
}
\author{
Fabian Mueller
}
