% Generated by roxygen2 (4.0.2): do not edit by hand
\name{theme_nogrid}
\alias{theme_nogrid}
\title{theme_nogrid}
\usage{
theme_nogrid(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{base size}

\item{base_family}{base family}
}
\value{
the theme structure
}
\description{
A ggplot2 theme based on theme_bw but with no grid lines and axis only on top and bottom
}
\examples{
theme_set(theme_nogrid())
dframe <- data.frame(x=runif(100),y=runif(100))
ggplot(dframe,aes(x=x,y=y)) + geom_point()
}

