% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimRed.R
\name{umapParamGridReport}
\alias{umapParamGridReport}
\title{umapParamGridReport}
\usage{
umapParamGridReport(X, outDir, metric = c("euclidean"),
  min_dist = c(0.01, 0.05, seq(0.1, 0.9, by = 0.1)), n_neighbors = c(5,
  15, 25, 50), ...)
}
\arguments{
\item{X}{feature matrix containing one row for each observation and one column for each feature}

\item{outDir}{output directory}

\item{metric}{parameters passed on to \code{uwot::umap()}}

\item{min_dist}{parameters passed on to \code{uwot::umap()}}

\item{n_neighbors}{parameters passed on to \code{uwot::umap()}}

\item{...}{parameters passed on to \code{getDimRedPlot}}

\item{colorBy}{annotation to color by}
}
\value{
a \code{muReportR} report (HTML) showing dimension reduction plots for the grid search
}
\description{
Generate a report with plots of UMAP dimension reduction plots for parameter combinations
}
\author{
Fabian Mueller
}
