% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringOps.R
\name{underscore2camel}
\alias{underscore2camel}
\title{underscore2camel}
\usage{
underscore2camel(x)
}
\arguments{
\item{a}{string or string vector}
}
\value{
the converted string(s)
}
\description{
converts underscores to camel case in a (vector of) string(s)
}
\examples{
underscore2camel("bla_blubb")
}
