% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{writeTab}
\alias{writeTab}
\title{writeTab}
\usage{
writeTab(fn, sep = "\\t", row.names = FALSE, col.names = TRUE,
  quote = FALSE, ...)
}
\arguments{
\item{fn}{filename to write to}

\item{sep}{see \code{?write.table}}

\item{row.names}{see \code{?write.table}}

\item{col.names}{see \code{?write.table}}

\item{quote}{see \code{?write.table}}

\item{...}{passed to \code{write.table}}

\item{x}{table to write to file}

\item{stringsAsFactors}{see \code{?write.table}}
}
\value{
the result of \code{write.table}
}
\description{
Wrapper around \code{write.table} to write tab-separated tables by default
}
\author{
Fabian Mueller
}
