% Generated by roxygen2 (4.0.2): do not edit by hand
\name{camel2underscore}
\alias{camel2underscore}
\title{camel2underscore}
\usage{
camel2underscore(x)
}
\arguments{
\item{a}{string or string vector}
}
\value{
the converted string(s)
}
\description{
converts camel case to underscores in a (vector of) string(s)
}
\examples{
underscore2camel("bla_blubb")
}

