% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{colpal.histone}
\alias{colgrad.methylation.rb}
\alias{colgrad.methylation.yb}
\alias{colpal.histone}
\alias{colpal.histone,colpal.histone.ihec,colgrad.methylation.rb,colgrad.methylation.yb}
\alias{colpal.histone.ihec}
\title{Custom Color Paletes for epigenetic modifications}
\format{\preformatted{ Named chr [1:10] "#7FCDBB" "#ADDD8E" "#238B45" "#FB6A4A" ...
 - attr(*, "names")= chr [1:10] "H3K4me1" "H3K4me2" "H3K4me3" "H3K9me1" ...
}}
\usage{
colpal.histone

colpal.histone.ihec

colgrad.methylation.rb

colgrad.methylation.yb
}
\description{
Named vectors of colors for different modifications
\describe{
  \item{\code{colpal.histone}}{
  	Histone modifications
  }
}

\describe{
  \item{\code{colpal.histone.ihec}}{
     Histone modifications. Colors defined by IHEC.
  }
}

\describe{
  \item{\code{colgrad.methylation.rb}}{
     3 point color palette for methylation: red-grey-blue
  }
}

\describe{
  \item{\code{colgrad.methylation.yb}}{
     3 point color palette for methylation: yellow-turquois-blue
  }
}
}
\examples{
library(gplots)
pie(rep(1,length(colpal.histone)), labels=names(colpal.histone), col=colpal.histone)
library(gplots)
pie(rep(1,length(colpal.histone.ihec)), labels=names(colpal.histone.ihec), col=colpal.histone.ihec)
library(plotrix)
cp <- colorpanel(100,colgrad.methylation.rb["low"],colgrad.methylation.rb["mid"],colgrad.methylation.rb["high"])
plot.new()
gradient.rect(0,0,1,1,col=cp,nslices=length(cp),gradient="x",border=NA)
library(plotrix)
cp <- colorpanel(100,colgrad.methylation.yb["low"],colgrad.methylation.yb["mid"],colgrad.methylation.yb["high"])
plot.new()
gradient.rect(0,0,1,1,col=cp,nslices=length(cp),gradient="x",border=NA)
}
\keyword{datasets}

