% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colpal.cb}
\alias{colpal.cb}
\alias{colpal,colpal.cb,colpal.colpal.bde}
\alias{colpal.bde}
\alias{colpal.nature}
\alias{colpal.mu.cat}
\alias{colpal.iwh.cb01}
\alias{colpal.solarextra}
\alias{colpal.corpid}
\alias{colpals.games}
\title{Custom Color Paletes}
\format{
An object of class \code{character} of length 9.

An object of class \code{character} of length 13.

An object of class \code{character} of length 17.

An object of class \code{character} of length 18.

An object of class \code{character} of length 23.

An object of class \code{character} of length 9.

An object of class \code{character} of length 13.

An object of class \code{list} of length 6.
}
\usage{
colpal.cb

colpal.bde

colpal.nature

colpal.mu.cat

colpal.iwh.cb01

colpal.solarextra

colpal.corpid

colpals.games
}
\description{
\describe{
  \item{\code{colpal.cb}}{
        color blind friendly color palettes (adapted from http://wiki.stdout.org/rcookbook/Graphs/Colors%20%28ggplot2%29/)
  }
}
}
\examples{
plotColpal(colpal.cb)
plotColpal(colpal.bde)
plotColpal(colpal.nature)
plotColpal(colpal.mu.cat)
plotColpal(colpal.iwh.cb01)
plotColpal(colpal.solarextra)
plotColpal(colpal.corpid)
plotColpal(colpals.games[["rollgalaxy"]])
}
\keyword{datasets}
