% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{colpal.histone.ihec}
\alias{colgrad.methylation.rb}
\alias{colgrad.methylation.yb}
\alias{colpal,colpal.cb,colpal.colpal.bde}
\alias{colpal.bde}
\alias{colpal.cb}
\alias{colpal.histone.ihec}
\title{\describe{
  \item{\code{colpal.histone.ihec}}{
     Histone modifications. Colors defined by IHEC.
  }
}}
\format{\preformatted{ Named chr [1:8] "#B2DF8A" "#1F78B4" "#A6CEE3" "#33A02C" ...
 - attr(*, "names")= chr [1:8] "H3K4me1" "H3K4me3" "H3K9me3" "H3K27me3" ...
}}
\usage{
colpal.histone.ihec

colgrad.methylation.rb

colgrad.methylation.yb

colpal.cb

colpal.bde
}
\description{
\describe{
  \item{\code{colpal.histone.ihec}}{
     Histone modifications. Colors defined by IHEC.
  }
}

\describe{
  \item{\code{colgrad.methylation.rb}}{
     3 point color palette for methylation: red-grey-blue
  }
}

\describe{
  \item{\code{colgrad.methylation.rb}}{
     3 point color palette for methylation: red-grey-blue
  }
}

\describe{
  \item{\code{colpal.cb}}{
        color blind friendly color palettes (adapted from http://wiki.stdout.org/rcookbook/Graphs/Colors%20%28ggplot2%29/)
  }
}

\describe{
  \item{\code{colpal.bde}}{
        Enhanced Color Brewer palette 'Dark2'
  }
}
}
\examples{
library(gplots)
pie(rep(1,length(colpal.histone.ihec)), labels=names(colpal.histone.ihec), col=colpal.histone.ihec)
library(plotrix)
cp <- colorpanel(100,colgrad.methylation.rb["low"],colgrad.methylation.rb["mid"],colgrad.methylation.rb["high"])
plot.new()
gradient.rect(0,0,1,1,col=cp,nslices=length(cp),gradient="x",border=NA)
library(plotrix)
cp <- colorpanel(100,colgrad.methylation.yb["low"],colgrad.methylation.yb["mid"],colgrad.methylation.yb["high"])
plot.new()
gradient.rect(0,0,1,1,col=cp,nslices=length(cp),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.cb,nslices=length(colpal.cb),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.bde,nslices=length(colpal.bde),gradient="x",border=NA)
}
\keyword{datasets}

