% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colpal.cb}
\alias{colpal.cb}
\alias{colpal,colpal.cb,colpal.colpal.bde}
\alias{colpal.bde}
\alias{colpal.nature}
\alias{colpal.mu.cat}
\alias{colpal.corpid}
\alias{colpals.games}
\title{Custom Color Paletes}
\format{An object of class \code{character} of length 9.}
\usage{
colpal.cb

colpal.bde

colpal.nature

colpal.mu.cat

colpal.corpid

colpals.games
}
\description{
\describe{
  \item{\code{colpal.cb}}{
        color blind friendly color palettes (adapted from http://wiki.stdout.org/rcookbook/Graphs/Colors%20%28ggplot2%29/)
  }
}

\describe{
  \item{\code{colpal.bde}}{
        Enhanced Color Brewer palette 'Dark2'
  }
}

\describe{
  \item{\code{colpal.nature}}{
        Color palette inspired by Nature journal color scheme's.
  }
}

\describe{
  \item{\code{colpal.nature}}{
        Color palette inspired by Nature journal color scheme's.
  }
}

\describe{
  \item{\code{colpal.corpid}}{
        Color palette inspired by coorporate identities I worked with
  }
}

\describe{
  \item{\code{colpals.games}}{
        List of color palette inspired by board games
  }
}
}
\examples{
plot.new()
gradient.rect(0,0,1,1,col=colpal.cb,nslices=length(colpal.cb),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.bde,nslices=length(colpal.bde),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.nature,nslices=length(colpal.nature),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.mu.cat,nslices=length(colpal.mu.cat),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.corpid,nslices=length(colpal.corpid),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpals.games[["bruges"]],nslices=length(colpals.games[["bruges"]]),gradient="x",border=NA)
}
\keyword{datasets}
