% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{colpal.cb}
\alias{colpal,colpal.cb,colpal.colpal.bde}
\alias{colpal.bde}
\alias{colpal.cb}
\title{Custom Color Paletes}
\format{\preformatted{ chr [1:9] "#E69F00" "#56B4E9" "#009E73" "#F0E442" "#0072B2" ...
}}
\usage{
colpal.cb

colpal.bde
}
\description{
\describe{
  \item{\code{colpal.cb}}{
        color blind friendly color palettes (adapted from http://wiki.stdout.org/rcookbook/Graphs/Colors%20%28ggplot2%29/)
  }
}

\describe{
  \item{\code{colpal.bde}}{
        Enhanced Color Brewer palette 'Dark2'
  }
}
}
\examples{
plot.new()
gradient.rect(0,0,1,1,col=colpal.cb,nslices=length(colpal.cb),gradient="x",border=NA)
plot.new()
gradient.rect(0,0,1,1,col=colpal.bde,nslices=length(colpal.bde),gradient="x",border=NA)
}
\keyword{datasets}

