% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getRegionSet}
\alias{getRegionSet}
\title{getRegionSet}
\usage{
getRegionSet(id, assembly = "hg19")
}
\arguments{
\item{id}{region set id}

\item{assembly}{genome assembly}
}
\value{
a GRanges object containing the region set
}
\description{
retrieves custom region sets by id
}
\examples{
regs <- getRegionSet("deep_chip_ctrl_regions")
}

