% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_ggplots2_ext.R
\name{ggsave4doc}
\alias{ggsave4doc}
\title{ggsave4doc}
\usage{
ggsave4doc(fn, plot = last_plot(), width = 146, height = 146,
  units = "mm", family = "Palatino", ...)
}
\arguments{
\item{fn}{file name}

\item{plot}{see \code{?ggsave}}

\item{width}{see \code{?ggsave}}

\item{height}{see \code{?ggsave}}

\item{units}{see \code{?ggsave}}

\item{family}{see \code{?ggsave}}

\item{plot}{see \code{?ggsave}}
}
\value{
result of \code{ggsave} command
}
\description{
Wrapper around ggsave that has default values for parameters fitting for embedding
plots into my documents
}
