% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{granges2bed}
\alias{granges2bed}
\title{granges2bed}
\usage{
granges2bed(gr, fn, score = NULL, addAnnotCols = FALSE,
  colNames = FALSE, doSort = TRUE, bigBed = FALSE,
  strandCharNA = ".")
}
\arguments{
\item{gr}{GRanges object}

\item{fn}{filename to save bed file to}

\item{addAnnotCols}{add the columns stored in elementMetadata of GRanges}

\item{colNames}{add column names}

\item{doSort}{sort the regions before writing the output}

\item{bigBed}{also save as bigbed file. Requires that the GRanges object has chromosome sizes stored.}

\item{strandCharNA}{character to be used if strand is NA, '*' or '.'}

\item{sc}{score vector or column in elementMetadata of GRanges}
}
\value{
(invisibly) the written results as a data.frame
}
\description{
Save a GRanges object to a bed file
}
