% Generated by roxygen2 (4.0.1): do not edit by hand
\name{normalize.str}
\alias{normalize.str}
\title{normalize.str}
\usage{
normalize.str(x, resolve.camel = FALSE, return.camel = FALSE)
}
\arguments{
\item{a}{string or string vector}

\item{resolve.camel}{Is the string in camelCase and should this be resolved to snake_case?}

\item{return.camel}{should camelCase be outputed rather than undescores?}
}
\value{
the normalized string(s)
}
\description{
normalize a string by removing special characters and
replacing whitespaces and dots and afterwards
remove all leading and trailing whitespaces and special characters
By default, underscore is the replacement character. Avoids consecutive underscores.
}
\examples{
normalize.str("_ (b)lA BLu[bb.blA\\tblubb- bla)\\n_")
}

