% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicRegions.R
\name{setGenomeProps}
\alias{setGenomeProps}
\title{setGenomeProps}
\usage{
setGenomeProps(gr, assembly, dropUnknownChrs = TRUE, adjChrNames = TRUE,
  ...)
}
\arguments{
\item{gr}{GRanges object or GAlignments object to modify}

\item{assembly}{assembly}

\item{dropUnknownChrs}{discard entries with seqnames not supported by assembly}

\item{adjChrNames}{should the prefix "chr" be added to main chromosomes if not already present and chrMT be renamed to chrM?}

\item{...}{arguments passed on to \code{getSeqlengths4assembly}}
}
\value{
GRanges object with genome properties set
}
\description{
Set the genome properties for a GRanges or GAlignments object given the name of a genome assembly
}
