% Generated by roxygen2 (4.0.1): do not edit by hand
\name{underscore2camel}
\alias{underscore2camel}
\title{underscore2camel}
\usage{
underscore2camel(x)
}
\arguments{
\item{a}{string or string vector}
}
\value{
the converted string(s)
}
\description{
converts underscores to camel case in a (vector of) string(s)
}
\examples{
underscore2camel("bla_blubb")
}

